+++ test/fixtures/object.js
